Got rid of warnings.
Former-commit-id: 1b954d0e3671079e232ad9814ba5acb2dfd8869f
Former-commit-id: ac0e56b09fad647b9fcce37a84cf9c51cc468272
Former-commit-id: 3441df78606bffbc6d825b2b5233259dfabc323c
Former-commit-id: 3f1cb5211f983fb6240b98867276f67e02887a06 [formerly b58e1bf28ef8442d8e5f142eb4ca768a12371fd7]
Former-commit-id: 5af6c791c0dd2c7e28e1bf7dfea450eb29500309

+++ src/tamm/eigensolver/src/getshifts.cpp
@@ -82,7 +82,7 @@ VectorXd getshifts(MPI_Comm comm, MatrixXd &H, MatrixXd &S, int nev, int nshifts
    int nPts = 5000;
    DOS.domain = linspace( evbnds(0), evbnds(1), nPts+2 );
    // Remove end points
-   for( int64_t j = 1; j < DOS.domain.size(); ++j )
+   for( size_t j = 1; j < DOS.domain.size(); ++j )
       DOS.domain[j-1] = DOS.domain[j];
    DOS.domain.pop_back();
    DOS.domain.pop_back();
